requirements.txt
Flask==0.9
Flask-SQLAlchemy==0.16
Jinja2==2.6
SQLAlchemy==0.8.0
Werkzeug==0.8.3
distribute==0.6.34
flask-heroku==0.1.4
gunicorn==0.17.2
psycopg2==2.5
py-bcrypt==0.3
wsgiref==0.1.2
